Move condition validation onto the conditions
Give each AccessCondition a Validate() alongside Evaluate() and remove the
IAccessConditionVisitor. AccessRuleValidator now delegates per-condition checks
to condition.Validate() and only enforces the document-level shape (array, size
limit, null-entry guard). TimeWindow gains its own Validate() plus a shared
TryParseTime, so Contains and validation use one HH:mm parse rather than a regex
that could drift from evaluation.

AccessRuleValidationResult moves into the conditions namespace so the condition
models do not depend on the service layer.
