+++ bitwarden_license/src/Services/Pam/Services/AccessRuleValidator.cs
@@ -1,5 +1,4 @@
-﻿using System.Net;
-using System.Text.Json;
+﻿using System.Text.Json;
 using Bit.Services.Pam.Models.Conditions;
 
 namespace Bit.Services.Pam.Services;
@@ -7,9 +6,6 @@ namespace Bit.Services.Pam.Services;
 public sealed class AccessRuleValidator : IAccessRuleValidator
 {
     private const int MaxConditions = 10;
-
-    // Stateless, so one shared instance serves every call.
-    private static readonly ConditionValidator _conditionValidator = new();
 
     public AccessRuleValidationResult Validate(string? conditionsJson)
     {
@@ -46,40 +42,15 @@ public sealed class AccessRuleValidator : IAccessRuleValidator
             return AccessRuleValidationResult.Invalid($"Conditions cannot contain more than {MaxConditions} conditions.");
         }
 
+        // Each condition validates itself; the validator only enforces the document-level shape (it is an array,
+        // within the size limit) and guards the one thing a condition cannot check for itself: a null entry.
         return conditions.Select(ValidateCondition).FirstOrDefault(result => !result.IsValid)
             ?? AccessRuleValidationResult.Valid;
     }
 
     private static AccessRuleValidationResult ValidateCondition(AccessCondition? condition) =>
+        // A JSON null in the array is not a condition and cannot validate itself, so it is rejected here.
         condition is null
             ? AccessRuleValidationResult.Invalid("Conditions cannot contain a null entry.")
-            : condition.Accept(_conditionValidator);
-
-    /// <summary>
-    /// Checks a single condition is well-formed. Stateless, mirroring the engine's evaluator: neither public
-    /// service is itself a visitor; each delegates to a private one.
-    /// </summary>
-    private sealed class ConditionValidator : IAccessConditionVisitor<AccessRuleValidationResult>
-    {
-        public AccessRuleValidationResult VisitHumanApproval(HumanApprovalCondition condition) =>
-            AccessRuleValidationResult.Valid;
-
-        public AccessRuleValidationResult VisitIpAllowlist(IpAllowlistCondition condition)
-        {
-            if (condition.Cidrs.Count == 0)
-            {
-                return AccessRuleValidationResult.Invalid("ip_allowlist requires at least one CIDR.");
-            }
-
-            foreach (var cidr in condition.Cidrs)
-            {
-                if (string.IsNullOrWhiteSpace(cidr) || !IPNetwork.TryParse(cidr, out _))
-                {
-                    return AccessRuleValidationResult.Invalid($"Invalid CIDR: '{cidr}'.");
-                }
-            }
-
-            return AccessRuleValidationResult.Valid;
-        }
-    }
+            : condition.Validate();
 }
