+++ bitwarden_license/src/Services/Pam/Services/IAccessRuleValidator.cs
@@ -1,4 +1,6 @@
-﻿namespace Bit.Services.Pam.Services;
+﻿using Bit.Services.Pam.Models.Conditions;
+
+namespace Bit.Services.Pam.Services;
 
 public interface IAccessRuleValidator
 {
@@ -7,10 +9,4 @@ public interface IAccessRuleValidator
     /// configured" and considered valid; callers decide how to treat that semantically.
     /// </summary>
     AccessRuleValidationResult Validate(string? conditionsJson);
-}
-
-public sealed record AccessRuleValidationResult(bool IsValid, string? Error)
-{
-    public static AccessRuleValidationResult Valid { get; } = new(true, null);
-    public static AccessRuleValidationResult Invalid(string error) => new(false, error);
 }
