+++ bitwarden_license/test/Services/Pam.Test/Commands/UpdateAccessRuleCommandTests.cs
@@ -3,6 +3,7 @@ using Bit.Core.Repositories;
 using Bit.Pam.Entities;
 using Bit.Pam.Models;
 using Bit.Pam.Repositories;
+using Bit.Services.Pam.Models.Conditions;
 using Bit.Services.Pam.OrganizationFeatures.Commands;
 using Bit.Services.Pam.Services;
 using Bit.Test.Common.AutoFixture;
