+++ bitwarden_license/test/Services/Pam.Test/Engine/AccessRuleEngineTests.cs
@@ -80,6 +80,6 @@ public class AccessRuleEngineTests
             return result;
         }
 
-        public override T Accept<T>(IAccessConditionVisitor<T> visitor) => throw new NotSupportedException();
+        public override AccessRuleValidationResult Validate() => AccessRuleValidationResult.Valid;
     }
 }
